+++ storage.go
@@ -1,0 +1,9 @@
+package main
+
+type Product struct {
+	ID    string  `json:"id"`
+	Name  string  `json:"name"`
+	Price float64 `json:"price"`
+}
+
+var Products []Product //In-memory data store