+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Next Release
 -
 
+* Products route with handlers
+  * Get -> returns all Products
+  * Post -> adds a new Product to the storage. Duplicates not allowed
+
 v0.0.0
 * Initial files and setup.